+++ spec/dependency_resolver_spec.rb
@@ -15,4 +15,18 @@ describe Dentaku::DependencyResolver do
       ["THIRD", "SeCoNd", "FIRST"]
     )
   end
+
+  it 'sorts mixed-case expressions in dependency order when case-sensitive' do
+    dependencies = {"TopLevel" => ["MidLevel"], "MidLevel" => ["BottomLevel"], "BottomLevel" => []}
+    expect(described_class.find_resolve_order(dependencies, true)).to eq(
+      ["BottomLevel", "MidLevel", "TopLevel"]
+    )
+  end
+
+  it 'treats variables that differ only in case as distinct when case-sensitive' do
+    dependencies = {"foo" => ["FOO"], "FOO" => []}
+    expect(described_class.find_resolve_order(dependencies, true)).to eq(
+      ["FOO", "foo"]
+    )
+  end
 end